Worked on Configure layout. Added Request object
The Request table will hold all of the actions that will
need to be sent to the RTM service at some point. The app
will cache these in case the request fails. The user can
also choose to have a "sync" time, or only sync manually.

The Configure layout will allow the user to set these
options, as well as present the Authenticate button to
allow them to authorize the app for syncing with RTM