Some error handling. Documentation.
I created a snazzy error dialog that can pop up to inform a user
that something has gone amiss. It includes a link to the issues
page of Github, so that they could potentially submit an ticket
if they have time. I also added some documentation to a few
methods.